Feature/tests init (#7)
* volume mount first test. 


* rename src to app, tests can import code modules now. 


* hot reloading, entrypoint one level higher for PYtest to be sidebyside. 


* pull routes into separate file. 


* separate threads from usage to allow patching. setup test client


* more reasonable files and tests. 


* not imported as module


* imports and better test.


* remove inconsistent test.


* thread sends post to agent test.


* set env as app.attribute.. is this bad? 


* Remove unused files from issuer controller.
Update docs to reflect standing up code in this repository.


* update deprecated warn method, remove debug line. 


* haven't actually used this yet. remove for now. 


* postman test suite.
